% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figures.R
\name{build_figures}
\alias{build_figures}
\title{Generates Figures for an \code{nlmixr} Report}
\usage{
build_figures(obnd = NULL, fit = NULL, rptdetails = NULL, verbose = TRUE)
}
\arguments{
\item{obnd}{onbrand report object to have report elements appended to}

\item{fit}{nlmixr fit object to be reported}

\item{rptdetails}{Object creating when reading in rptyaml file}

\item{verbose}{Boolean variable when set to TRUE (default) messages will be
displayed on the terminal}
}
\value{
List containing the figures with the following structure:
\itemize{
\item \code{"rptfigs"} - List of figures with names corresponding to the
figure ids in the yaml file. It contains the elements from the yamle file
and the following elements:
\itemize{
\item \code{"figure"}       - ggplot object or the name of the file containing the figure
\item \code{"ftype"}        - Type of figure either "ggplot" or "imagefile"
\item \code{"isgood"}       - Boolean variable indicating success or failure
\item \code{"fmsgs"}        - Vector of messages
\item \code{"cmd_proc"}     - Plot generation command after processing for placeholders
\item \code{"caption_proc"} - Caption after processing for placeholders
}
\item \code{"isgood"} - Boolean variable indicating success or failure
\item \code{"msgs"} - Vector of messages
}
}
\description{
Creates figures specified in a rptyaml file
}
