% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{eval_str}
\alias{eval_str}
\title{Evaluate R Code in String}
\usage{
eval_str(estr = "", fit = NULL)
}
\arguments{
\item{estr}{Object creating when reading in rptyaml file}

\item{fit}{nlmixr2 fit object to be reported}
}
\value{
String containing the evaled as a character or the original string
}
\description{
Attempts to evaluate a string as a chunk of R code.
}
\examples{
res = eval_str(estr="ls()")
}
