% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{fetch_fit_example}
\alias{fetch_fit_example}
\title{Fetch Fit Example}
\usage{
fetch_fit_example()
}
\value{
Example nlmixr2 fit object
}
\description{
Creates an nlmixr2 fit example using \code{posthoc} estimation
method for testing purposes.
displayed on the terminal
following:
}
\examples{
fit = fetch_fit_example()
}
