% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{fetch_option}
\alias{fetch_option}
\title{Fetch Analysis Options}
\usage{
fetch_option(rptdetails, option = NULL, fit = NULL, verbose = TRUE)
}
\arguments{
\item{rptdetails}{Object creating when reading in rptyaml file}

\item{option}{String containing the option to fetch (see below)}

\item{fit}{nlmixr fit object to be reported}

\item{verbose}{Boolean variable when set to \code{TRUE} (default) messages will be
displayed on the terminal
following:}
}
\value{
List containing the following information about the output directory
\itemize{
\item \code{"isgood"} - Boolean variable indicating success or failure
\item \code{"msgs"} - Vector of messages
\item \code{"value"} - The value of the option or the default if not
specified
}
}
\description{
Fetches analysis options from the report yaml
applies it to strings.
}
\details{
The option can be one of the following (default: \code{NULL}):
\itemize{
\item \code{"output_dir"} - Directory to place figures that are generated (default: \code{tempdir()})
\item \code{"resolution"} - Resolution of figure files (default: \code{300})
\item \code{"length_units"} - Units (default: \code{"in"})
}
}
