% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tables.R
\name{mk_error_tab}
\alias{mk_error_tab}
\title{Generates a \code{flextable} Object with Error Message}
\usage{
mk_error_tab(msgs)
}
\arguments{
\item{msgs}{Vector of error messages}
}
\value{
list with a single \code{flextable} object
}
\description{
Takes a vector of messages and returns a \code{flextable} object with the
text in the table. This can be used in automated table generation to
cascade an error message to the end user.
}
\examples{
error_tab = mk_error_tab("This is an error")
error_tab$ft[[1]]
}
