% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{process_ph}
\alias{process_ph}
\title{Substitutes Placehodlers in Strings}
\usage{
process_ph(str, rptdetails)
}
\arguments{
\item{str}{String to process}

\item{rptdetails}{Object creating when reading in rptyaml file
(default: \code{NULL})}
}
\value{
processed string
}
\description{
Takes placeholder information from the rptyaml file and
applies it to strings.
}
