% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{report_fit}
\alias{report_fit}
\title{Report \code{nlmixr} Fit Results to PowerPoint and Word}
\usage{
report_fit(
  obnd = NULL,
  fit = NULL,
  rptyaml = system.file(package = "rptnlmixr", "templates", "report_fit.yaml")
)
}
\arguments{
\item{obnd}{onbrand report object to have report elements appended to}

\item{fit}{nlmixr fit object to be reported}

\item{rptyaml}{yaml file containing the report elements and structure}
}
\value{
onbrand object with the report elements added
}
\description{
Appends \code{nlmixr} fit results to an onbrand report object with the
content and format of the report in the supplied yaml file
}
