% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{safe_text}
\alias{safe_text}
\title{Escapes Curly Braces for cli functions}
\usage{
safe_text(x)
}
\arguments{
\item{x}{String with curly braces}
}
\value{
String with curly braces escaped
}
\description{
Converts curly braces to double curly braces in order to prevent interpertation with the cli package
}
\examples{
safe_text("This string has curly {braces}")
}
