% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_figures.R
\name{write_figure}
\alias{write_figure}
\title{Gets Figure Dimensions}
\usage{
write_figure(
  p_res = NULL,
  page = NULL,
  width = 3,
  height = 3,
  resolution = NULL,
  fig_file = NULL,
  verbose = TRUE
)
}
\arguments{
\item{p_res}{ggplot or ggforce paginated object}

\item{page}{page number to write or NULL for a ggplot object}

\item{width}{width in inches}

\item{height}{height in inches}

\item{resolution}{resolution in dpi}

\item{fig_file}{file name to write the figure to}

\item{verbose}{Boolean variable when set to TRUE (default) messages will be
displayed on the terminal}
}
\value{
list with the following
\itemize{
\item \code{"isgood"} - Boolean variable indicating success or failure
\item \code{"msgs"} - Vector of messages
}
}
\description{
For a given figure id and report type this will pull out the
dimensions of the figure.
}
