% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tables.R
\name{build_tables}
\alias{build_tables}
\title{Generates Tables  for an \code{nlmixr} Report}
\usage{
build_tables(obnd = NULL, fit = NULL, rptdetails = NULL, verbose = TRUE)
}
\arguments{
\item{obnd}{onbrand report object to have report elements appended to}

\item{fit}{nlmixr fit object to be reported}

\item{rptdetails}{object creating when reading in rptyaml file}

\item{verbose}{Boolean variable when set to TRUE (default) messages will be
displayed on the terminal}
}
\value{
list of tables objects with names taken from the tables element in
the rptdetails object
}
\description{
Creates tables specified in a rptyaml file
}
