% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rptnlmixr.R
\name{report_fit}
\alias{report_fit}
\title{Report \code{nlmixr2} Fit Results to PowerPoint and Word}
\usage{
report_fit(
  obnd = NULL,
  fit = NULL,
  rptyaml = system.file(package = "nlmixr2rpt", "templates", "report_fit.yaml"),
  ph = NULL
)
}
\arguments{
\item{obnd}{onbrand report object to have report elements appended to}

\item{fit}{nlmixr2 fit object to be reported}

\item{rptyaml}{yaml file containing the report elements and structure}

\item{ph}{Manual placeholders, see \code{\link{yaml_read_fit}} for more
details
for example RUN may be "RUNN" in the yaml file. To overwrite this just
provide \code{list(RUN="RUN_1")} (default: \code{NULL})}
}
\value{
onbrand object with the report elements added
}
\description{
Appends \code{nlmixr2} fit results to an onbrand report object with the
content and format of the report in the supplied yaml file
}
